LayoutTests/crypto/subtle/ec-import-spki-key-export-spki-key-p384-expected.txt

Test importing a P-384 SPKI ECDH key and exporting it in SPKI format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS bytesToHexString(publicKey) is spkiKeyHex
PASS successfullyParsed is true

TEST COMPLETE
